# openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Intelligence Report Generator API",
    "version": "1.0.0",
    "description": "Generate military-grade intelligence assessment PDF reports with satellite imagery, threat matrices, scenario analysis, and priority actions. Supports multi-domain threat assessment, vulnerability analysis, predictive outlook, and data quality confidence tables. Reports include classification banners, On Demand branding, and professional formatting."
  },
  "servers": [
    {
      "url": "https://YOUR_DEPLOYED_URL"
    }
  ],
  "paths": {
    "/intelligence-report/start": {
      "post": {
        "operationId": "startSession",
        "summary": "Start a new intelligence report session",
        "description": "Initialize a new session to begin building an intelligence report. Returns a sessionId to use in subsequent calls.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sessionId"],
                "properties": {
                  "sessionId": {
                    "type": "string",
                    "description": "A unique session identifier for this report (e.g., 'uae-iran-2026-03-07')"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "description": "Success message" },
                    "sessionId": { "type": "string", "description": "The session ID" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/intelligence-report/generate": {
      "post": {
        "operationId": "generateReport",
        "summary": "Generate the intelligence report PDF",
        "description": "Compile all session data into a professionally formatted PDF intelligence report and return a download URL. The session is cleaned up after generation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sessionId"],
                "properties": {
                  "sessionId": {
                    "type": "string",
                    "description": "The session ID to generate the report for"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "description": "Success message" },
                    "url": { "type": "string", "description": "URL to download the generated PDF report" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}